Move sidebar component into dir along with sidebar container.
