+++ src/sidebar-common/components/index.tsx
@@ -1,12 +1,11 @@
-export { default as AnnotationBox } from './annotation-box'
+export { default as AnnotationBoxContainer } from './annotation-box'
 export { default as AnnotationHighlight } from './annotation-highlight'
 export { default as CloseButton } from './close-button'
 export { default as CongratsMessage } from './congrats-message'
 export { default as EmptyMessage } from './empty-message'
 export { default as Footer } from './footer'
 export { default as Loader } from './loader'
-export { default as Sidebar } from './sidebar'
 export { default as TagHolder } from './tag-holder'
 export { default as TagInput } from './tag-input'
-export { default as TopBar } from './topbar'
+export { default as Topbar } from './topbar'
 export { default as TruncatedTextRenderer } from './truncated-text-renderer'
