--- src/sidebar-common/components/sidebar/index.tsx
@@ -1,3 +1,0 @@
-import Sidebar from './sidebar'
-
-export default Sidebar
