+++ src/sidebar-common/sidebar/components/index.tsx
@@ -1,0 +1,3 @@
+import SidebarContainer from './sidebar-container'
+
+export default SidebarContainer
