+++ src/sidebar-common/sidebar/components/menu-styles.ts
