+++ src/sidebar-common/sidebar/components/sidebar.css
