+++ src/sidebar-common/sidebar/components/sidebar.tsx
@@ -1,14 +1,16 @@
 import * as React from 'react'
 import Menu from 'react-burger-menu/lib/menus/slide'
 
-import CongratsMessage from '../congrats-message'
+import {
+    CongratsMessage,
+    Topbar,
+    Loader,
+    EmptyMessage,
+    AnnotationBoxContainer,
+} from '../../components'
 import menuStyles from './menu-styles'
 import CommentBoxContainer from '../../comment-box'
-import Topbar from '../topbar'
-import Loader from '../loader'
-import EmptyMessage from '../empty-message'
-import AnnotationBoxContainer from '../annotation-box'
-import { Annotation } from '../../sidebar/types'
+import { Annotation } from '../types'
 import { openSettings } from '../../utils'
 
 const styles = require('./sidebar.css')
